refactor: import paths for types & refactor paths for sia functions to be in one place
