+++ components/dashboard/assets/assetsTable.tsx
@@ -18,7 +18,7 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import React from "react";
-import { TAsset, TConversionRates } from "@/lib/types";
+import { TAsset, TConversionRates } from "@/types/types";
 import AssetsTableCard from "@/components/mobile/assetsTableCard";
 
 interface DataTableProps<TValue> {
